feat(ops): migrate reconcile + duplicates to v2 OperationDefs
Replace all 10 s.queue.Enqueue calls in reconcile.go and
duplicates_handlers.go with s.opRegistry.EnqueueOp, registering
corresponding OperationDefs in new reconcile_ops.go and
duplicates_ops.go files. v1 CreateOperation calls are kept for
backward compat. All guards updated from s.queue==nil to
s.opRegistry==nil.
